geometry update
