+++ imap_processing/spice/geometry.py
@@ -14,6 +14,7 @@ import typing
 from enum import IntEnum
 from pathlib import Path
 from typing import Union
+from numpy.typing import NDArray
 
 import numpy as np
 import numpy.typing as npt
@@ -303,6 +304,8 @@ def get_spacecraft_to_instrument_spin_phase_offset(instrument: SpiceFrame) -> fl
     return offset_lookup[instrument]
 
 
+@typing.no_type_check
+@ensure_spice
 def frame_transform(
     et: Union[float, npt.NDArray],
     position: npt.NDArray,
@@ -368,8 +371,6 @@ def frame_transform(
     return result
 
 
-@typing.no_type_check
-@ensure_spice
 def get_rotation_matrix(
     et: Union[float, npt.NDArray],
     from_frame: SpiceFrame,
@@ -489,3 +490,75 @@ def basis_vectors(
     ... spacecraft_z = basis_vectors[:, 2]
     """
     return np.moveaxis(get_rotation_matrix(et, from_frame, to_frame), -1, -2)
+
+
+def cartesian_to_spherical(
+    v: NDArray,
+) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Convert cartesian coordinates to spherical coordinates.
+
+    Parameters
+    ----------
+    v : np.ndarray
+        A NumPy array with shape (n, 3) where each
+        row represents a vector
+        with x, y, z-components.
+
+    Returns
+    -------
+    az : np.ndarray
+        The azimuth angles in degrees.
+    el : np.ndarray
+        The elevation angles in degrees.
+    r : np.ndarray
+        The radii, or magnitudes, of the vectors.
+    """
+    vx = v[:, 0]
+    vy = v[:, 1]
+    vz = v[:, 2]
+
+    # Magnitude of the velocity vector
+    magnitude_v = np.sqrt(vx**2 + vy**2 + vz**2)
+
+    vhat_x = -vx / magnitude_v
+    vhat_y = -vy / magnitude_v
+    vhat_z = -vz / magnitude_v
+
+    # Elevation angle (angle from the z-axis, range: [-pi/2, pi/2])
+    el = np.arcsin(vhat_z)
+
+    # Azimuth angle (angle in the xy-plane, range: [0, 2*pi])
+    az = np.arctan2(vhat_y, vhat_x)
+
+    # Ensure azimuth is from 0 to 2PI
+    az = az % (2 * np.pi)
+
+    return np.degrees(az), np.degrees(el), magnitude_v
+
+
+def spherical_to_cartesian(
+    r: np.ndarray, theta: np.ndarray, phi: np.ndarray
+) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+    """
+    Convert spherical coordinates to Cartesian coordinates.
+
+    Parameters
+    ----------
+    r : np.ndarray
+        Radius.
+    theta : np.ndarray
+        Azimuth angle in radians.
+    phi : array-like or float
+        Elevation angle in radians.
+
+    Returns
+    -------
+    x, y, z : tuple
+        Cartesian coordinates.
+    """
+    x = r * np.cos(phi) * np.cos(theta)
+    y = r * np.cos(phi) * np.sin(theta)
+    z = r * np.sin(phi)
+
+    return x, y, z
